Sun, Mar 6, 2022, 12:35 PM +00:00

+++ readability/readability.c
@@ -15,11 +15,11 @@ int main(void)
     int sentences = count_sentences(txt);
     float L = letters / (float)words * 100;
     float S = sentences / (float)words * 100;
-    float index = 0.0588 * L - 0.296 * S - 15.8;
+    int index = 0.0588 * L - 0.296 * S - 15.8;
     printf("%i letters\n", letters);
     printf("%i words\n", words + 1);
     printf("%i sentences\n", sentences);
-    printf("%f index\n", index);
+    printf("%i index\n", index);
 }
 
 int count_letters(string text)
